requirements.txt
art==5.4
coloredlogs==15.0.1
djangorestframework==3.13.1
elasticsearch==7.17.0
emoji==1.6.3
firebase-admin==5.2.0
google-cloud-storage==2.0.0
pydantic==1.9.0
pymongo==4.0.1
python-slugify==5.0.2
redis==4.1.3
requests==2.27.1
